styling and frontend to hookup Delete and Update
