+++ web/src/pages/vehicles.js
@@ -7,22 +7,18 @@ import DataStore from '../util/DataStore';
 /**
  * Logic needed for the view vehicle page of the website.
  */
-class ViewVehicle extends BindingClass {
+class Vehicles extends BindingClass {
   constructor() {
     super();
-    this.bindClassMethods(['mount', 'getVehicleForPage', 'getRecordsForPage',
-                           'createNewRecord','updateRecordsTable'], this);
+    this.bindClassMethods(['mount','getVehicleForPage'], this);
     this.dataStore = new DataStore();
     this.header = new Header(this.dataStore);
-    this.dataStore.addChangeListener(this.updateRecordsTable);
   }
 
   mount() {
     this.header.addHeaderToPage();
     this.client = new WrenchWenchClient();
     this.getVehicleForPage();
-    this.getRecordsForPage();
-    document.getElementById('submitNewRecordButton').addEventListener('click', this.createNewRecord);
   }
 
   async getVehicleForPage() {
@@ -44,7 +40,7 @@ class ViewVehicle extends BindingClass {
 
       let html =
            `<div class="table-responsive-xl">
-                <table class="table">
+                <table class="table-striped">
                   <tr>
                     <th>VIN</th>
                     <th>Make</th>
@@ -130,93 +126,28 @@ class ViewVehicle extends BindingClass {
            vehicleTableContainer.innerHTML += html;
   }
 
-  async getRecordsForPage() {
 
-    const urlParams = new URLSearchParams(window.location.search);
-    const vin = urlParams.get('vin');
-    const records = await this.client.getVehicleRecordsList(vin);
+  async submit(evt) {
+          evt.preventDefault();
 
-    if (records !== null) {
-      this.dataStore.set('records', records);
-      console.log('Records have been stored.');
-    } else {
-      console.log('Records were not found');
-    }
-  }
+          const vin = document.getElementById('inputVin').value; // Get the value from the "inputVin" input field
+          const vehicles = await this.client.createVehicle(vin);
 
-  updateRecordsTable() {
-    const records = this.dataStore.get('records');
-    if (records === null) {
-      return;
-    }
+          this.dataStore.set('vehicles', vehicles);
+       }
 
-    const recordsTable = document.getElementById('records-table');
-    recordsTable.innerHTML = ''; // Clear the table
-
-
-
-    records.recordsList.forEach((record) => {
-      const recordId = record.recordId;
-      const description = record.description;
-      const status = record.status;
-      const priorityLevel = record.priorityLevel;
-      const timestamp = record.timestamp;
-      const isNewRecord = record.isNewRecord
-
-      const recordModel = record.recordModel;
-      console.log('Record in UpdateRecordTable: ', record); // Debugging statement
-
-      let html = `<tr>`
-
-      if(isNewRecord){
-        html += `<td>${recordModel.recordId}</td><td>${recordModel.description}</td>` +
-        `<td>${recordModel.status}</td><td>${recordModel.priorityLevel}</td><td>${recordModel.timestamp}</td>`;
-
-      } else {
-        html += `<td>${record.recordId}</td><td>${description}</td>` +
-        `<td>${status}</td><td>${priorityLevel}</td><td>${timestamp}</td>`;
+      redirectToViewVehicle(){
+          const vehicles = this.dataStore.get('vehicles');
+          const vin = document.getElementById('inputVin').value;
+          if(vehicles != null){
+              window.location.href = `/viewVehicle.html?vin=${vehicles.vin}`;
+          }
       }
-      html += `</tr>`;
-      console.log('HTML: ', html);
-
-      recordsTable.innerHTML += html;
-    });
-  }
-
-  async createNewRecord(evt) {
-    const urlParams = new URLSearchParams(window.location.search);
-    const vin = urlParams.get('vin');
-    evt.preventDefault();
-
-    const checkedValue = $("input[name='flexRadioDefault']:checked").attr('value');
-
-    const description = document.getElementById('input-description').value;
-    const priorityLevel = checkedValue;
-
-    const record = await this.client.createVehicleRecord(vin, description, priorityLevel);
-
-    if (record) {
-        record.isNewRecord = true;
-
-        this.dataStore.update((state) => {
-          const updatedRecords = { ...state.records};
-          updatedRecords.recordsList.push(record);
-          return { ...state, records: updatedRecords};
-        });
-        console.log('Record in CreateNewRecord: ', record); // Debugging statement
-        console.log('New Record Saved');
-    } else {
-        console.log('Failed to create a new record');
-    }
-  }
 }
 
 const main = async () => {
-  const viewVehicle = new ViewVehicle();
-  viewVehicle.mount();
-  $('#myCollapsible').collapse({
-    toggle: false,
-  });
+  const vehicles = new Vehicles();
+  vehicles.mount();
 };
 
 window.addEventListener('DOMContentLoaded', main);
