+++ web/static_assets/css/style.css
@@ -134,6 +134,7 @@ https://developer.mozilla.org/en-US/docs/Web/CSS/CSS_Lists_and_Counters/Using_CS
 th, td {
     padding: 15px;
     text-align: left;
+
 }
 
 th {
