--- web/static_assets/garage.html
@@ -1,40 +1,0 @@
-<!DOCTYPE html>
-<html>
-    <head>
-        <script type="text/javascript" src="assets/loadGarage.js"></script>
-        <link rel="stylesheet" type="text/css" href="css/style.css">
-    </head>
-
-    <body class="background">
-        <header class="header" id="header">
-            <button  type="button" class="button" id="logout">Login</button>
-        </header>
-
-        <div class="garage-list">
-            <h2>Garage Table</h2>
-
-            <div id="getVehicle">
-                <form id="getVehicleForm" class="card-content">
-                    <p class="form-field">
-                        <label>Search by name or tag</label>
-                        <input type="text" class="validated-field" id="vin-input" placeholder="smooth-jamz" autofocus>
-                        <sub>Search is case-sensitive and tags must match exactly.</sub>
-                    </p>
-                </form>
-            </div>
-
-            <p class="button-group">
-                <a href="#" class="button" id="search-btn">Search</a>
-                <a href="viewVehicles.html" class="button">Vehicles List</a>
-            </p>
-        </div>
-        <div class="card hidden" id='maintenance-records-results-container'>
-
-            <h3>Results for <span id="records-criteria-display"></span></h3>
-            <div id="maintenance-records-results-display">
-            </div>
-        </div>
-
-    </body>
-
-</html>
