+++ web/static_assets/index.html
@@ -4,7 +4,7 @@
     <head>
         <meta charset="utf-8">
         <meta name ="viewport" content="width=device-width, initial-scale=1">
-        <script type="text/javascript" src="assets/createVehicle.js"></script>
+        <script type="text/javascript" src="assets/index.js"></script>
         <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-9ndCyUaIbzAi2FUVXJi0CjmCapSmO7SnpJef0486qhLnuZ2cdeRhO02iuK6FUUVM" crossorigin="anonymous">
         <title>Wrench Wench</title>
     </head>
@@ -38,48 +38,19 @@
         <div class="px-3 py-2 bg-dark text-white">
             <div class="container">
 
-                <div class="d-flex flex-wrap align-items-center justify-content-center justify-content-lg-start">
+                <div class="d-flex flex-wrap align-items-start justify-content-md-start justify-content-lg-start">
                     <a href="/" class="d-flex align-items-center my-2 my-lg-0 me-lg-auto text-white text-decoration-none">
                         <svg class="bi me-2" width="40" height="32" role="img" aria-label="Bootstrap"><use xlink:href="#bootstrap"></use></svg>
                     </a>
 
-                    <ul class="nav col-12 col-lg-auto my-2 justify-content-center my-md-0 text-small">
+                    <ul class="nav col-12 col-lg-auto my-2 justify-content-md-start my-md-0 text-small">
                         <li>
                             <a href="#" class="nav-link text-secondary">
                                 <svg class="bi d-block mx-auto mb-1" width="24" height="24"><use xlink:href="#home"></use></svg>
                                 Home
                             </a>
                         </li>
-                        <li>
-                            <a href="#" class="nav-link text-white">
-                                <svg class="bi d-block mx-auto mb-1" width="24" height="24"><use xlink:href="#tools"></use></svg>
-                                Vehicles
-                            </a>
-                        </li>
-                        <li>
-                            <a href="#" class="nav-link text-white">
-                                <svg class="bi d-block mx-auto mb-1" width="24" height="24"><use xlink:href="#table"></use></svg>
-                                Records
-                            </a>
-                        </li>
-                        <li>
-                            <a href="#" class="nav-link text-white">
-                                <svg class="bi d-block mx-auto mb-1" width="24" height="24"><use xlink:href="#grid"></use></svg>
-                                Products
-                            </a>
-                        </li>
-                        <li>
-                            <a href="#" class="nav-link text-white">
-                                <svg class="bi d-block mx-auto mb-1" width="24" height="24"><use xlink:href="#people-circle"></use></svg>
-                                Customers
-                            </a>
-                        </li>
                     </ul>
-
-                    <div class="text-end">
-                        <button type="button" id="login" class="btn btn-light text-dark me-2">Login</button>
-                        <button type="button" class="btn btn-primary">Sign-up</button>
-                    </div>
                 </div>
             </div>
         </div>
@@ -93,11 +64,10 @@
             <form class="col-12 col-lg-auto mb-2 mb-lg-0 me-lg-auto">
                 <input type="text" id="inputVin" class="form-control" placeholder="Must be exactly 17 characters" aria-label="inputVin" required>
             </form>
-
-
+            <button class="btn btn-secondary" id="submitVinButton">Submit</button>
 
         </div>
-        <button class="btn btn-primary" id="submitVinButton">Submit</button>
+
 
     </div>
