--- web/static_assets/vehicle.html
@@ -1,16 +1,0 @@
-<!DOCTYPE html>
-<html>
-
-<head>
-    <script type="text/javascript" src="assets/viewVehicle.js"></script>
-    <link rel="stylesheet" type="text/css" href="css/style.css">
-</head>
-
-<body class="background">
-    <header class="header" id="header"></header>
-
-
-
-</body>
-
-</html>
